Extract scalar aggregation rewriter to a separate class
This code will be reused in the iterative optimizer
